leaf71: add ifnum else support v0
