+++ crates/carreltex-engine/src/compile_v0/macro_expand_v0.rs
@@ -1,12 +1,12 @@
 use std::collections::BTreeMap;
 
+use super::ifnum_v0::parse_ifnum_v0;
 use crate::reasons_v0::InvalidInputReasonV0;
 use crate::tex::tokenize_v0::{TokenV0, MAX_TOKENS_V0};
 
 pub(crate) const MAX_MACROS_V0: usize = 4096;
 pub(crate) const MAX_MACRO_EXPANSIONS_V0: usize = 4096;
 pub(crate) const MAX_MACRO_DEPTH_V0: usize = 64;
-pub(crate) const MAX_IF_DEPTH_V0: usize = 64;
 const MAX_COUNT_VALUE_V0: u32 = 1_000_000;
 
 #[derive(Clone)]
@@ -158,6 +158,9 @@ fn expand_stream_v0(
                     depth + 1,
                 )?;
                 index = next_index;
+            }
+            TokenV0::ControlSeq(name) if name.as_slice() == b"else" => {
+                return Err(InvalidInputReasonV0::MacroIfElseWithoutIf);
             }
             TokenV0::ControlSeq(name) => {
                 match lookup_macro_binding_v0(macro_frames, name) {
@@ -432,89 +435,6 @@ fn parse_noexpand_v0(
         .clone();
     push_checked_v0(out, next)?;
     Ok(noexpand_index + 2)
-}
-
-fn parse_ifnum_v0(
-    tokens: &[TokenV0],
-    ifnum_index: usize,
-    counters: &[u32; 2],
-    if_depth: usize,
-) -> Result<(Vec<TokenV0>, usize), InvalidInputReasonV0> {
-    if if_depth >= MAX_IF_DEPTH_V0 {
-        return Err(InvalidInputReasonV0::MacroIfDepthExceeded);
-    }
-
-    let mut index = skip_space_tokens_v0(tokens, ifnum_index + 1);
-    let left = parse_ifnum_count_operand_v0(tokens, &mut index, counters)?;
-    index = skip_space_tokens_v0(tokens, index);
-    let operator = match tokens.get(index) {
-        Some(TokenV0::Char(b'<')) => b'<',
-        Some(TokenV0::Char(b'=')) => b'=',
-        Some(TokenV0::Char(b'>')) => b'>',
-        _ => return Err(InvalidInputReasonV0::MacroIfnumUnsupported),
-    };
-    index += 1;
-    index = skip_space_tokens_v0(tokens, index);
-    let right = parse_ifnum_count_operand_v0(tokens, &mut index, counters)?;
-    let condition = match operator {
-        b'<' => left < right,
-        b'=' => left == right,
-        b'>' => left > right,
-        _ => return Err(InvalidInputReasonV0::MacroIfnumUnsupported),
-    };
-
-    let mut out = Vec::<TokenV0>::new();
-    while index < tokens.len() {
-        match tokens.get(index) {
-            Some(TokenV0::ControlSeq(name)) if name.as_slice() == b"ifnum" => {
-                let (nested_tokens, next_index) = parse_ifnum_v0(tokens, index, counters, if_depth + 1)?;
-                if condition {
-                    for token in nested_tokens {
-                        push_checked_v0(&mut out, token)?;
-                    }
-                }
-                index = next_index;
-            }
-            Some(TokenV0::ControlSeq(name)) if name.as_slice() == b"fi" => {
-                return Ok((out, index + 1));
-            }
-            Some(TokenV0::ControlSeq(name)) if name.as_slice() == b"else" => {
-                return Err(InvalidInputReasonV0::MacroIfnumUnsupported);
-            }
-            Some(token) => {
-                if condition {
-                    push_checked_v0(&mut out, token.clone())?;
-                }
-                index += 1;
-            }
-            None => break,
-        }
-    }
-
-    Err(InvalidInputReasonV0::MacroIfnumUnsupported)
-}
-
-fn parse_ifnum_count_operand_v0(
-    tokens: &[TokenV0],
-    index: &mut usize,
-    counters: &[u32; 2],
-) -> Result<u32, InvalidInputReasonV0> {
-    if !matches!(tokens.get(*index), Some(TokenV0::ControlSeq(name)) if name.as_slice() == b"count")
-    {
-        return Err(InvalidInputReasonV0::MacroIfnumUnsupported);
-    }
-    *index += 1;
-    match tokens.get(*index) {
-        Some(TokenV0::Char(b'0')) => {
-            *index += 1;
-            Ok(counters[0])
-        }
-        Some(TokenV0::Char(b'1')) => {
-            *index += 1;
-            Ok(counters[1])
-        }
-        _ => Err(InvalidInputReasonV0::MacroIfnumUnsupported),
-    }
 }
 
 fn parse_let_v0(
